+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `merchantAccountId` parameter to Transaction refund
+
 ## 5.5.0
 * Add `scaExemption` to Transaction sale
 * Deprecate `deviceSessionId` and `fraudMerchantId` in `CreditCardGateway`, `CustomerGateway`, `PaymentMethodGateway`, and `TransactionGateway` classes